feat : fitpet->slimpet 패키지 명 변경
