+++ android/app/build.gradle
@@ -87,9 +87,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.fitpet"
+    namespace "com.slimpet"
     defaultConfig {
-        applicationId "com.fitpet"
+        applicationId "com.slimpet"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
