+++ android/app/src/main/java/com/slimpet/StepSensorModule.kt
@@ -1,0 +1,50 @@
+package com.slimpet
+
+import android.content.Context
+import android.hardware.Sensor
+import android.hardware.SensorEvent
+import android.hardware.SensorEventListener
+import android.hardware.SensorManager
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContextBaseJavaModule
+import com.facebook.react.bridge.ReactMethod
+import com.facebook.react.modules.core.DeviceEventManagerModule
+
+class StepSensorModule(private val reactContext: ReactApplicationContext) :
+    ReactContextBaseJavaModule(reactContext), SensorEventListener {
+
+    private val sensorManager =
+        reactContext.getSystemService(Context.SENSOR_SERVICE) as SensorManager
+
+    override fun getName() = "StepSensorModule"
+
+    @ReactMethod
+    fun startListening() {
+        val sensor = sensorManager.getDefaultSensor(Sensor.TYPE_STEP_COUNTER)
+        if (sensor != null) {
+            sensorManager.registerListener(this, sensor, SensorManager.SENSOR_DELAY_FASTEST)
+        }
+    }
+
+    @ReactMethod
+    fun stopListening() {
+        sensorManager.unregisterListener(this)
+    }
+
+    override fun onSensorChanged(event: SensorEvent) {
+        if (event.sensor.type == Sensor.TYPE_STEP_COUNTER) {
+            val totalSinceBoot = event.values[0].toDouble()
+            reactContext
+                .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+                .emit("StepCounterUpdate", totalSinceBoot)
+        }
+    }
+
+    override fun onAccuracyChanged(sensor: Sensor?, accuracy: Int) {}
+
+    @ReactMethod
+    fun addListener(eventName: String) {}
+
+    @ReactMethod
+    fun removeListeners(count: Int) {}
+}
