+++ ios/SlimPet/Images.xcassets/AppIcon.appiconset/Icon-1024.png
