+++ ios/SlimPet/Images.xcassets/AppIcon.appiconset/Icon-20@2x.png
