+++ ios/SlimPet/Images.xcassets/AppIcon.appiconset/Icon-60@2x.png
